feat: #455 suport batch gulp.init
When in batch mode, instead of calling gulp.init, initialize the store by hand.

This is not the optimal way to do this because if gulp.init changes, the batch
mode code may become obsolete; but given that there's no easy way to make batch
mode configuration available to gulp tasks, I prefer to have a much simpler
though more error prone code than a very complex but perfect one. Especially
since we will need to revisit these things in the future to align JS and Themes
Toolkit and hopefully will find a way to fix this limitation.