chore: move disable-publish to prepublishOnly
The disable-publish fails the publish unless run from a release script to avoid
developers doing `yarn publish`. This was inspired by the publish-please
package.

Problem is that it should be called in prepublishOnly, as prepublish is invoked
by `yarn add` too and thus makes it fail.

Note, however, that this will be removed in the future when we integrate
liferay-js-publish in the project.