feat: add no-arrow rule (#179)
To be applied selectively where we don't want arrow functions to break
IE when we deliver untranspiled code.

Not a breaking change because it isn't turned on by default.

There are more sophisticated tools out there like:

https://github.com/mysticatea/eslint-plugin-es

Which have autofixes, but we want this to be a temporary measure anyway
(the real fix is transpiling our themes JS) so let's not bring in
another dependency.

Closes: https://github.com/liferay/eslint-config-liferay/issues/179