feat(bundler): enhance error handling
Move ExplainedError to its own file and stop using it in imports-loader.

Instead, log parse errors in imports-loader to the report file and don't fail
the build.

Note that if we've been unable to parse the file it's because it is not valid
JavaScript so there shouldn't be any import in there. Also, it is almost certain
that webpack will fail after us when parsing the same content, so we leave the
error handling to webpack better to avoid user confusion.