feat(bundler): add 'source' configuration value
We need a source dir configuration folder for projects where the source files
are contained in a nested directory inside the project (like it's the case of
Liferay Portal).

This is because we must strip the path to the nested folder before using file
paths as module names and things like that.