feat: #230 enhance resolution errors handling
Instead of sending resolution errors inside fields which are not intended for
that (resolvedModules, moduleMap and so on) create new warnings and errors
fields to hold them.

The loader will then read those fields and, if anything is found, the messages
will be dumped to the console with the necessary log level.

Also, if any error (not warning) is sent from the server, the resolution will
fail.