fix: #584 add interop for imports
In situations where a __webpack_require__ is transformed into a
Liferay.Loader.require because an import has been configured, there are sometimes
issues with the ES6 default exports because no interop code is executed.

For example, we have this code in the portal (frontend-taglib module):

var ClayTabs = __REQUIRE__('@frontend-taglib-clay$clayui/tabs');

that fetches ClayTabs from an external imported module in frontend-taglib-clay.
As it is shown, this import gets the verbatim object (without indexing the
default field) when it should pass it through the interop code.