fix: give up on trying to check codepen.io links
Since:

    https://github.com/liferay/liferay-frontend-guidelines/pull/161

Cloudflare has been 503-ing these links.

They used to 403 them, and we got around that in:

    https://github.com/liferay/liferay-frontend-guidelines/pull/134

But it looks like the party is over now, and I don't think it is worth
trying to keep up with them on this one. Another example of a PR with a
spurious failure due to these links:

    https://github.com/liferay/liferay-frontend-guidelines/pull/163