fix: #588 enhance error detection on dynamic requires
NOTE: THIS DOES NOT FIX #588 ENTIRELY, JUST MAKES IT A BIT LESS PROBLEMATIC

I have added code to ignore dynamic requires during build time. Now we don't
generate the undefined dependency, just skip it and let it fail during execution
of the module (note that it may even not fail if the code flow doesn't make the
dynamic require to be executed).

Also, an error is written to the report file.

This should alleviate the problem because:

1. It may make some setups work (as long as they don't execute the dynamic
   require)
2. If they execute the require, a better error will be shown in the console
3. The problem is correctly explained in the report file