fix(scripts): catch whole-script indentation problems in JSP (#467)
The way our JS-in-JSP formatting works is as follows:

1. Extract JS.
2. Pass through Prettier.
3. Put JS back where it was.

This means that if the entire block is misindented (ie. too far to the
left or the right) but each line is correctly indented relative to its
neighbors, the formatting won't complain, because the round trip via
Prettier won't actually change anything.

So, we need add an additional layer of enforcement here, which is that
the initial indent is correct compared to its containing block.

The fix is straightforward, but I also wanted to make sure that it
played nicely in an edge case scenario like this:

    <script>alert('boom');</script>

We don't have any tags like that in liferay-portal, but if we did we
wouldn't want to mess them up. In general, the Java source formatter
wants opening and closing tags to be on lines of their own. So I made it
reformat them to:

    <script>
        alert('boom');
    </script>

Closes: https://github.com/liferay/liferay-npm-tools/issues/467