refactor(scripts): move a comment
To where it may be more helpful in explaining how this works. Also,
gives me a commit message in which I can provide some detail if somebody
`git-blame`s it.

Given input like this (showing tabs as visible with "\t"):

    \t  \t  <aui:script>
    \t  \t  \t  hey();
    ...

The JS block start (`range.start.column`) will be 15 (ie. 2 tabs, plus
the 12 characters of "<aui:script>", plus 1 because columns are
1-indexed).

Assuming the Java source formatter has done the right thing (and we can
fairly assume that), we can deduce that the base indent of the script
contents should be 3 tabs without actually having to look for and count
tabs; ie.

    range.start.column (15) - "<aui:script>".length (12)