docs(dxp): swap screenshots and clarify
Whoops. I had the screenshots back to front, so swap them.

Also, make it super explicit where the "allow edits" checkbox is; sadly,
I don't think there's an account-level or repo-level setting for this,
but it is a "sticky" setting so doesn't need to be set every time. Not
sure if it is "sticky" if you use the `gh` tool, but I guess we will
find out.