refactor: #606 extract jar creation to script
We don't want the bundler to be responsible for the JAR assembly. This is
because in Java based environments, this assembly is done by gradle, for
example.

However, some people may still need JARs to be created without having to install
a JDK, so we will still support that scenario with the
liferay-js-toolkit-scripts package.