fix(changelog): use prefixes for non-monorepos too
Noticed while looking at:

    https://github.com/liferay/liferay-ckeditor/releases

That the links to the "full changelog" are missing the "v" prefix in a
lot of places.

That's because the code we have to read the prefix from the ".yarnrc"
was inadequately tested outside of monorepos. This small fix makes
sure that it reads the top-level ".yarnrc" as well, for the non-monorepo
case.

Test plan: Make a merge commit in liferay-ckeditor and run the
generator with and without a "v" prefix. Then apply this change and
repeat the test; results:

1. "master", with "v" prefix: get valid links in changelog.
2. "master", without "v" prefix: get bad links in changelog.
3. This commit, with "v" prefix: get valid links in changelog.
4. This commit, without "v" prefix: get valid links in changelog.

ie. broken case "2" is fixed.