feat: #625 bundle fragment projects
We'll be using bundler to prepare React based fragment projects.

It would be great if we could bundle them as regular bundler projects, but they have
some subtle differences that need to be accounted for when building them.

Basically we need to make sure that:

1) We only create one bundle per fragment (thus
we must disable webpack optimization and forget about entry points or extra
modules being generated).
2) We only import 'react' from the portal (because any other import will fail as
fragments don't have an associated OSGi bundle from which to resolve dependency
version numbers).
3) We don't need to tweak package.json or manifest.json as they won't be used.
4) We need to copy fragment assets so that fragments CLI can find them.