feat(css-loader): use preload when available (#631)
As described in the related ticket, the idea is to have the css-loader
emit:

    <link as="style" rel="preload" />

tags instead of:

    <link rel="stylesheet" />

ones.

As noted in:

    https://www.smashingmagazine.com/2016/02/preload-what-is-it-good-for/

> On top of that, preload **does not block the window’s `onload`
> event**, unless the resource is also requested by a resource that
> blocks that event.

In practice, we don't expect this to make page loads perceptibly faster,
but it should at least avoid Lighthouse reporting onload events as
delayed.

The `try`/`catch` is so that we can do feature detection for
the "preload" attribute and only use it on browsers that support it.
Without this, the `onload` event may not fire on IE which would lead to
breakage.

See: https://stackoverflow.com/questions/49136399/onload-on-link-element-event-doesnt-fire-on-ie11-preload-fallback-for-ie

Related: https://issues.liferay.com/browse/LPS-115876

Related: https://liferay-community.slack.com/archives/CB2QQ3M5F/p1598598237007500

Closes: https://github.com/liferay/liferay-js-toolkit/issues/631