chore: update JS toolkit dependencies from v2.18.4 to v2.19.0
Primarily to get this PR:

    https://github.com/liferay/liferay-js-toolkit/pull/632
    ("use preload when in link tags when available")

From the 2.19.0 release:

    https://github.com/liferay/liferay-js-toolkit/releases/tag/v2.19.0

Looking in the releases since v2.18.4 which we skipped and never
integrated into liferay-portal, I don't think there is anything
else that will really affect us, as they were mostly prepared for
outside-of-liferay-portal use cases:

- v2.18.5: https://github.com/liferay/liferay-js-toolkit/milestone/55?closed=1
- v2.18.6: https://github.com/liferay/liferay-js-toolkit/milestone/57?closed=1
- v2.18.7: https://github.com/liferay/liferay-js-toolkit/milestone/59?closed=1
- v2.18.8: https://github.com/liferay/liferay-js-toolkit/milestone/60?closed=1

One odd thing to note: when I did this update; ie. using:

    cd packages/liferay-npm-bundler-preset-liferay-dev
    yarn add \
      babel-preset-liferay-standard@2.19.0 \
      liferay-npm-bundler-loader-css-loader@2.19.0 \
      liferay-npm-bundler-plugin-{exclude-imports,inject-imports-dependencies,inject-peer-dependencies,namespace-packages,replace-browser-modules,resolve-linked-dependencies}@2.19.0
    cd ../liferay-npm-scripts
    yarn add \
      liferay-npm-bridge-generator@2.19.0 \
      liferay-npm-bundler@2.19.0

I got duplicate versions in the yarn.lock, even though `yarn why` would
only report a single version of each package, and `npx yarn-deduplicate
yarn.lock` did not remove the duplicates. I am not sure if this is
related to me updating from yarn v1.22.4 to v1.22.5 this morning, but in
any case, I was able to get the duplication to go away by running `yarn`
from the top-level (just running it, not deleting the lockfile
beforehand). I don't recall it behaving like this at any time in the
past.