fix: make namespace-amd-define work with css-loader
As noted here:

    https://github.com/liferay-frontend/liferay-portal/pull/345#issuecomment-684867451

Code like this which is produced by the css-loader:

```js
function defineModule() {
  define("frontend-js-web@4.0.27/liferay/modal/Modal.scss", ['module', 'exports', 'require'], function (module, exports, require) {
    var define = undefined;
    var global = window;
    {
      module.exports = link;
    }
  });
}
```

winds up with a naked `define` in it, instead of the expected
`Liferay.Loader.define`.

This behavior regressed in 0c14c6970503d66d73cb5b311d15ce5748d9eca5.

Now, I don't have enough context to know whether this is the right fix,
because the issues listed in that commit are kind of opaque to me.
Hopefully @izaera can confirm or refute that the proposed change here is
safe.

Of course, it's not very clever; if you have code like this:

```js
function a() {
  define([], function () {});
}

define([], function () {});
```

It is going to namespace the first `define` it sees and warn about the
second one, but it's totally arbitrary; there is nothing "special" about
the first one that appeared higher up in the file. If move the function
to the bottom, then the other `define` "wins" and gets transformed.