docs: add initial naming guidelines
This came up in code review again:

https://github.com/liferay/clay/pull/3684#discussion_r484662169

So I figure it is time to finally document this.

For now starting with abbreviations, but we can add more naming-related
things in the future. It is one of the two hard problems, after all...

The examples of "bad" abbreviations are all taken from past code
reviews.

Not intended to be exhaustive advice, but it's a start.

Closes: https://github.com/liferay/liferay-frontend-guidelines/issues/159