revert: feat(css-loader): use preload when available (#631)
This reverts commit 1423ca31e3a1f10, because the feature doesn't work as
expected, as described here:

https://github.com/liferay-frontend/liferay-portal/pull/345#issuecomment-694221607

In short, our testing showed the loader producing code that correctly
inserted `preload` links in the `head`, and the network pane showed
the corresponding resources being fetched, and in the right order. But
the associated styles were not being applied to the DOM, leading to
visual inconsistencies, as well as undesirable console spew of the form:

> The resource .../css/ApplicationsMenu.css was preloaded using link
> preload but not used within a few seconds from the window's load
> event. Please make sure it has an appropriate `as` value and it is
> preloaded intentionally.

To fix this, we might be able to add a second, non-`preload`-ing `link`
tag, but that would defeat the purpose of our original change, which was
to improve our Lighthouse metrics.

So, reverting this for now. We might revisit `preload` for this or some
other purpose in the future.

Original PR: https://github.com/liferay/liferay-js-toolkit/pull/632