feat(scripts): add testing-library/react-hooks
We were using v1.0.4 unofficially in liferay-portal due to this
`devDependency`:

https://github.com/liferay/liferay-portal/blob/f86d976d9f59933928199ee8b57ebee1c88efbd7/modules/dxp/apps/portal-workflow/portal-workflow-metrics-web/package.json#L36

We want to make it official and integrate it into liferay-portal "The
Right Way™", which means bring it into liferay-npm-scripts. The question
is, which version? The latest release is in the v3 range, which means
there have been two breaking changes.

In this commit, we try out the latest and we'll see what breaks. During
the install `yarn` prints:

    warning " > @testing-library/react-hooks@3.4.1" has unmet peer dependency "react@>=16.9.0".
    warning " > @testing-library/react-hooks@3.4.1" has unmet peer dependency "react-test-renderer@>=16.9.0".

First one is ok because we'll get that from the copy hoisted from
frontend-js-react-web in liferay-portal anyway (we have a very
loose requirement specified in liferay-npm-scripts' package.json so that
we don't wind up pulling in another version by mistake).

Second will be ok too because I am adding it as an explicit dep; again
with a loose version requirement because we want yarn to have maximum
flexibility to resolve the dependency to a single, unique version.

Test plan:

1. `yarn add` this in liferay-portal.
2. Grab https://github.com/liferay-frontend/liferay-portal/pull/384 and
   edit it; see that tests still work.

Closes: https://github.com/liferay/liferay-npm-tools/issues/496