feat: add support/filter-tags.sh
Which I am going to use in the next commit to rewrite unnamespaced tags
(eg. "v1.0.0") to use project namespaces (eg. "some-project/v1.0.0").

It's a specialized version of the `--tag-filter` functionality in
`git-filter-branch`. Ideally we'd just use `git-filter-branch`, like
this:

    git filter-branch \
      --commit-filter 'echo $GIT_COMMIT' \
      --tag-name-filter 'sed "s#^v#some-project/v#"'

but sadly, we can't do that. The idea of the `--commit-filter` is to be
a no-op, because we don't actually want `git-filter-branch` to rewrite
the commits (which will break the hashes and ruin internal links in
commit messages that use SHA-1 hashes). Unfortunately, it doesn't work
because `git-filter-branch` freaks out at the end of the run when we
wind up with the same hash as we started.

So, we use this custom bash script to just rewrite the tags and nothing
else. I'll show how to use it when I make the next commit, which will
merge in the history from the eslint-config-liferay repository.