refactor: migrate liferay-changelog-generator to @liferay/changelog-generator
Note that the binary continues to be called
`liferay-changelog-generator` because we want the package to use a named
scope but that doesn't mean that we want executable names to change due
to the way the `$PATH` gets searched. I believe `npx
@liferay/changelog-generator` will work, however, although I can't test
that until I cut a test release, which I will do shortly.