refactor: migrate liferay-npm-scripts to @liferay/npm-scripts
Only "tricky" bit here is updating the "contrib" scripts to make them
work with both `@liferay/npm-scripts` (which will likely continue to be
active on older DXP branches for a while) and `@liferay/npm-scripts`
(which will take over on `master` real soon now).

Also note one link in MultiTester.js which I had to loosen because I
won't have a concrete hash to link to until after I've created the this
commit.