docs: move CI badge up to top-level README
As I've noted in other commits, for now we are just going to run CI over
everything in the monorepo. Once that gets too big/slow we will look at
making the runs "smarter" (eg. only run on projects which have changes
in them).

But for now, given that the run is global, the badge belongs somewhere
global too.